Docs: Update expected output to show np.float64

+++ package/MDAnalysis/analysis/rms.py
@@ -243,8 +243,8 @@ def rmsd(a, b, weights=None, center=False, superposition=False):
     >>> A = bb.positions.copy()  # coordinates of first frame
     >>> _ = u.trajectory[-1]  # forward to last frame
     >>> B = bb.positions.copy()  # coordinates of last frame
-    >>> float(rmsd(A, B, center=True))
-    6.838544558398293
+    >>> rmsd(A, B, center=True)
+    np.float64(6.838544558398293)
 
 
     .. versionchanged:: 0.8.1
